Explain Use of Git to Manage Character Inventory
Explain to user the purpose of the folder, which is
designed to allow the user the ability to track inventory
using git whilst following the story. Then explain what
the user needs to know about commits, branches, merges,
stashes, etc. to allow for adequate tracking. Note that
the following files of the inventory: clothes.md,
weapons.md, and weather.md are often subject to change and
thus create the potential for merge conflicts.